[standalone] created berrybeam top wrapper
The berrybeam wrapper handles the operating mode of the program.
It sets the cfg.APP_MODE variable which is read by the child modes.

It follows an arg/subargs structure so that every mode can have its
parameter set, although in the future we might move it to a yaml.