[standalone] standalone mode sketch
It is not tested yet, but this is a skeleton of what it will look like.
Every mode execution is independent, meaning it spawns its own thread and
nrf24 config.

When a new mode is set via GPIO switches, the standalone handler changes
APP_MODE to the new value (which terminates the last mode) and spawns a
new thread with the new mode.

All the GPIO is also managed in the standalone function.